requirements.txt
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
astroid==2.8.6
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
biszx-pylint-odoo==1.0.6
black==22.1.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.2
django-crispy-forms==1.14.0
djangorestframework==3.13.1
docutils==0.17.1
entrypoints==0.3
executing==0.8.2
filelock==3.4.2
flake8==4.0.1
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
fonttools==4.29.1
identify==2.4.12
idna==3.3
importlib-metadata==4.11.3
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
nodeenv==1.6.0
notebook==6.4.8
numpy==1.22.1
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2022.1.8
platformdirs==2.4.1
polib==1.1.1
pre-commit==2.17.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==4.21.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.10.0
pylint==2.11.1
pylint-odoo==5.0.5
pylint-plugin-utils==0.6
pymessenger==0.0.7.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-toolbelt==0.9.1
restructuredtext-lint==1.3.2
rfc3986==2.0.0
roots-tools==1.1.4
Send2Trash==1.8.0
setuptools-odoo==3.0.7
setuptools-scm==6.4.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
stack-data==0.1.4
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
whichcraft==0.6.1
wrapt==1.13.3
zipp==3.7.0
